+++ merkle-interval-tree/Cargo.toml
@@ -8,4 +8,5 @@ edition = "2018"
 rust-crypto = {git="https://github.com/debris/rust-crypto.git", branch="android"}
 bytes = "0.4.12"
 byteorder = "1.3.1"
+ethereum-types = "^0.5.2"
 num-traits = { version = "0.2.8", default-features = false }
